Explicit cast to avoid warning.
git-svn-id: https://svn.aros.org/svn/aros/trunk@14782 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/exec/childwait.c
@@ -96,7 +96,7 @@
 	/* Check if it has returned already. This will also take the first. */
 	ForeachNode(&et->et_TaskMsgPort.mp_MsgList, child)
 	{
-	    if (tid == 0 || child->et_UniqueID == tid)
+	    if ((ULONG)tid == 0 || child->et_UniqueID == (ULONG)tid)
 		break;
 	}
